pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
alembic = "^1.9.4"
SQLAlchemy = {extras = ["mypy", "asyncio"], version = "^2.0.4"}
httpx = "^0.23.0"
uvicorn = "^0.19.0"
uvloop = "^0.17.0"
asyncpg = "^0.27.0"
loguru = "^0.6.0"
pydantic = {extras = ["email", "dotenv"], version = "^1.10.5"}
aiofiles = "^22.1.0"
sentry-sdk = { extras = ["fastapi"], version = "^1.10.1" }
python-multipart = "^0.0.5"
celery = { extras = ["redis"], version = "^5.2.7" }
passlib = {version="^1.7.4", extras=["bcrypt"]}
pendulum = "^2.1.2"
anyio = "^3.6.2"
prometheus-client = "^0.15.0"
orjson = "^3.8.1"
httptools = "^0.5.0"
redis = { extras = ["hiredis"], version = "^4.4.0rc2" }
attrs = "^22.1.0"
python-docx = "^0.8.11"
gunicorn = "^20.1.0"
uuid7 = "^0.1.0"
pillow = "^10.0.0"
python-json-logger = "^2.0.7"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
python-jose = "^3.3.0"
flower = "^2.0.1"
jinja2 = "^3.1.2"
pyhumps = "^3.8.0"


[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
cyclonedx-bom = "^3.11.0"
pytest-sugar = "^0.9.6"
pytest-asyncio = "^0.20.1"
coverage = "^6.5.0"
autoflake = "1.4"
black = "^22.6.0"
flake8 = "3.9.2"
isort = "5.8.0"
mypy = "^0.991"
pyinstrument = "^4.1.1"
pika-stubs = "^0.1.3"
types-redis = "^4.3.21.4"
ruff = "^0.0.265"
bandit = { extras = ["toml"], version = "^1.7.5"}
pytest-cov = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
line_length = 88
force_grid_wrap = 0
combine_as_imports = true
src_paths=["app","tests"]

[tool.black]
line-length = 88

[tool.bandit]
exclude_dirs = ["migrations", ".venv"]
tests = ["B201", "B301"]
skips = ["B101", "B601"]

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.11.
target-version = "py311"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10